+++ CHANGES.rst
@@ -57,8 +57,9 @@ Development
 - Prettify logging
 - Optionally start modeserver, defaulting to true
 - Add basic HTTP API subsystem. Thanks `@vkuhlen`_ and `@DieDiren`_.
-- Add ``make wifi-ap`` command for starting the access point interactively
+- Add ``make provide-wifi`` command for starting the access point interactively
 - Start WiFi in STA_AP mode by default
+- Disable garbage collector when reading sensors to improve timing
 
 
 2019-06-22 0.5.1